Break widgets apart into submodules
