+++ ryven/nodes/pyiron/atomistics_nodes.py
@@ -7,7 +7,7 @@ import numpy as np
 from pyiron_atomistics import Project
 from pyiron_atomistics.lammps import list_potentials
 from ryven.NENV import Node, NodeInputBP, NodeOutputBP, dtypes
-from ryven.ironflow.canvas.widgets import RepresentableNodeWidget, ButtonNodeWidget
+from ryven.ironflow.canvas.nodes import RepresentableNodeWidget, ButtonNodeWidget
 
 from abc import ABC, abstractmethod
 from ryven.nodes.std.special_nodes import DualNodeBase
